begin moving session context out of default context
