+++ src/vs/workbench/api/common/positron/extHost.positron.api.impl.ts
@@ -36,7 +36,6 @@ import { IToolInvocationContext } from '../../../contrib/chat/common/languageMod
 import { IPositronLanguageModelSource } from '../../../contrib/positronAssistant/common/interfaces/positronAssistantService.js';
 import { ExtHostEnvironment } from './extHostEnvironment.js';
 import { ExtHostPlotsService } from './extHostPlotsService.js';
-import { URI } from '../../../../base/common/uri.js';
 
 /**
  * Factory interface for creating an instance of the Positron API.
@@ -267,16 +266,7 @@ export function createPositronApiFactoryAndRegisterActors(accessor: ServicesAcce
 				return extHostAiFeatures.responseProgress(context, part);
 			},
 			getPositronChatContext(request: vscode.ChatRequest): Thenable<positron.ai.ChatContext> {
-				return extHostAiFeatures.getPositronChatContext(request)
-					.then((context) => {
-						return {
-							...context,
-							activeSession: context.activeSession && {
-								...context.activeSession,
-								notebookUri: context.activeSession.notebookUri && URI.revive(context.activeSession.notebookUri),
-							},
-						};
-					});
+				return extHostAiFeatures.getPositronChatContext(request);
 			},
 			getSupportedProviders(): Thenable<string[]> {
 				return extHostAiFeatures.getSupportedProviders();
