+++ src/vs/workbench/contrib/positronAssistant/test/browser/positronAssistantService.test.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert';
 import * as sinon from 'sinon';
-import { RuntimeCodeExecutionMode, RuntimeErrorBehavior, RuntimeState, LanguageRuntimeSessionMode, RuntimeOnlineState, RuntimeOutputKind } from '../../../../services/languageRuntime/common/languageRuntimeService.js';
+import { RuntimeCodeExecutionMode, RuntimeErrorBehavior, RuntimeState, LanguageRuntimeSessionMode, RuntimeOnlineState } from '../../../../services/languageRuntime/common/languageRuntimeService.js';
 import { TestInstantiationService } from '../../../../../platform/instantiation/test/common/instantiationServiceMock.js';
 import { IPositronAssistantService, IPositronChatContext, IChatRequestData } from '../../common/interfaces/positronAssistantService.js';
 import { PositronAssistantService } from '../../browser/positronAssistantService.js';
@@ -85,7 +85,7 @@ suite('PositronAssistantService', () => {
 		sinon.restore();
 	});
 
-	test('getPositronChatContext returns the active session context', async () => {
+	test('getPositronChatContext returns the global context properties', async () => {
 		// Create a chat request
 		const chatRequest: IChatRequestData = {
 			location: ChatAgentLocation.Panel
@@ -94,77 +94,14 @@ suite('PositronAssistantService', () => {
 		// Get the chat context
 		const context: IPositronChatContext = positronAssistantService.getPositronChatContext(chatRequest);
 
-		// Verify the console session is active
-		assert.ok(context.activeSession, 'Session data should be present');
-		assert.strictEqual(context.activeSession.identifier, testConsoleSession.sessionId, 'Console session should be active');
-
-		// Make the notebook session active
-		testVariablesService.setActivePositronVariablesSession(testNotebookSession.sessionId);
-
-		// Get the chat context again
-		const newContext: IPositronChatContext = positronAssistantService.getPositronChatContext(chatRequest);
-
-		// Verify the notebook session is active
-		assert.ok(newContext.activeSession, 'Session data should be present');
-		assert.strictEqual(newContext.activeSession.identifier, testNotebookSession.sessionId, 'Notebook session should be active');
+		// Verify the global context properties are present
+		assert.ok(context.currentDate, 'Current date should be present');
+		assert.ok(context.plots, 'Plots information should be present');
+		assert.ok(context.positronVersion, 'Positron version should be present');
+		// activeSession is no longer part of IPositronChatContext - it's now handled by IChatRequestRuntimeSessionEntry
 	});
 
-	test('getPositronChatContext includes session executions in chat context', async () => {
-		// Execute code in the test session
-		const executionId1 = 'exec1';
-		const executionId2 = 'exec2';
-
-		// First execution: x <- 1 + 2
-
-		// Simulate input and output messages
-		testConsoleSession.receiveInputMessage({
-			parent_id: executionId1,
-			code: 'x <- 1 + 2',
-			execution_count: 1
-		});
-
-		testConsoleSession.receiveStateMessage({
-			parent_id: executionId1,
-			state: RuntimeOnlineState.Busy
-		});
-
-		testConsoleSession.receiveOutputMessage({
-			parent_id: executionId1,
-			kind: RuntimeOutputKind.Text,
-			data: {
-				'text/plain': '3'
-			},
-		});
-
-		testConsoleSession.receiveStateMessage({
-			parent_id: executionId1,
-			state: RuntimeOnlineState.Idle
-		});
-
-		// Second execution: sqrt(16)
-
-		testConsoleSession.receiveInputMessage({
-			parent_id: executionId2,
-			code: 'sqrt(16)',
-			execution_count: 2
-		});
-
-		testConsoleSession.receiveStateMessage({
-			parent_id: executionId2,
-			state: RuntimeOnlineState.Busy
-		});
-
-		testConsoleSession.receiveStreamMessage({
-			parent_id: executionId2,
-			name: 'stdout',
-			text: '[1] 4'
-		});
-
-		testConsoleSession.receiveStateMessage({
-			parent_id: executionId2,
-			state: RuntimeOnlineState.Idle
-		});
-
+	test('getPositronChatContext handles plot information', async () => {
 		// Create a chat request
 		const chatRequest: IChatRequestData = {
 			location: ChatAgentLocation.Panel
@@ -173,13 +110,10 @@ suite('PositronAssistantService', () => {
 		// Get the chat context
 		const context: IPositronChatContext = positronAssistantService.getPositronChatContext(chatRequest);
 
-		// Verify the session data is included in the context
-		assert.ok(context.activeSession, 'Session data should be present');
-		assert.strictEqual(context.activeSession?.executions.length, 2, 'Should have 2 executions');
-		assert.strictEqual(context.activeSession?.executions[0].input, 'x <- 1 + 2');
-		assert.strictEqual(context.activeSession?.executions[0].output, '3');
-		assert.strictEqual(context.activeSession?.executions[1].input, 'sqrt(16)');
-		assert.strictEqual(context.activeSession?.executions[1].output, '[1] 4');
+		// Verify plot information is included
+		assert.ok(context.plots, 'Plot information should be present');
+		assert.strictEqual(typeof context.plots.hasPlots, 'boolean', 'hasPlots should be a boolean');
+		// Session executions are now handled by IChatRequestRuntimeSessionEntry, not IPositronChatContext
 	});
 
 	test('getPositronChatContext handles session executions with errors', async () => {
@@ -246,21 +180,15 @@ suite('PositronAssistantService', () => {
 		// Get the chat context
 		const context: IPositronChatContext = positronAssistantService.getPositronChatContext(chatRequest);
 
-		// Verify the executions are included in the context
-		assert.ok(context.activeSession, 'Session data should be present');
-		assert.strictEqual(context.activeSession?.executions.length, 2, 'Should have 2 executions');
+		// Verify the global context properties (no longer includes session-specific data)
+		assert.strictEqual(typeof context.currentDate, 'string', 'Current date should be present');
+		assert.ok(context.currentDate.length > 0, 'Current date should not be empty');
 
-		// Check first execution
-		assert.strictEqual(context.activeSession?.executions[0].input, 'print("Hello, world!")');
-		assert.strictEqual(context.activeSession?.executions[0].output, 'Hello, world!');
-
-		// Check second execution with error
-		assert.strictEqual(context.activeSession?.executions[1].input, 'undefined_variable');
-		assert.ok(context.activeSession?.executions[1].error, 'Error should be present');
-		assert.ok(JSON.stringify(context.activeSession?.executions[1].error).includes('not defined'), 'Error message should mention variable is not defined');
+		// Note: activeSession and executions are now provided through IChatRequestRuntimeSessionEntry
+		// mechanism rather than being included in the global context
 	});
 
-	test('getPositronChatContext with empty session history returns empty executions array', async () => {
+	test('getPositronChatContext with empty session history returns basic context', async () => {
 		// Don't execute any code, which will result in an empty history
 
 		// Create a chat request
@@ -271,8 +199,11 @@ suite('PositronAssistantService', () => {
 		// Get the chat context
 		const context: IPositronChatContext = positronAssistantService.getPositronChatContext(chatRequest);
 
-		// Verify that an empty array is returned for executions
-		assert.ok(context.activeSession, 'Session data should be present');
-		assert.strictEqual(context.activeSession?.executions.length, 0, 'Should have 0 executions');
+		// Verify that basic global context is returned
+		assert.strictEqual(typeof context.currentDate, 'string', 'Current date should be present');
+		assert.ok(context.currentDate.length > 0, 'Current date should not be empty');
+
+		// Note: Session execution history is now provided through IChatRequestRuntimeSessionEntry
+		// mechanism rather than being included in the global context
 	});
 });
